refactor: Move LoadingHelper to its own file
